refactor(taraxa_capability): TaraxaPeer class moved to the standalone header
